+++ photutils/psf/epsf.py
@@ -18,6 +18,7 @@ import numpy as np
 
 from .epsf_stars import EPSFStar, EPSFStars, LinkedEPSFStar
 from .models import EPSFModel
+from .utils import _interpolate_missing_data
 from ..centroids import centroid_com
 from ..utils._optional_deps import HAS_BOTTLENECK, HAS_TQDM  # noqa
 from ..utils._parameters import as_pair
@@ -817,59 +818,3 @@ class EPSFBuilder:
             pbar.close()
 
         return epsf, stars
-
-
-def _interpolate_missing_data(data, mask, method='cubic'):
-    """
-    Interpolate missing data as identified by the ``mask`` keyword.
-
-    Parameters
-    ----------
-    data : 2D `~numpy.ndarray`
-        An array containing the 2D image.
-
-    mask : 2D bool `~numpy.ndarray`
-        A 2D boolean mask array with the same shape as the input
-        ``data``, where a `True` value indicates the corresponding
-        element of ``data`` is masked.  The masked data points are
-        those that will be interpolated.
-
-    method : {'cubic', 'nearest'}, optional
-        The method of used to interpolate the missing data:
-
-        * ``'cubic'``:  Masked data are interpolated using 2D cubic
-            splines.  This is the default.
-
-        * ``'nearest'``:  Masked data are interpolated using
-            nearest-neighbor interpolation.
-
-    Returns
-    -------
-    data_interp : 2D `~numpy.ndarray`
-        The interpolated 2D image.
-    """
-    from scipy import interpolate
-
-    data_interp = np.array(data, copy=True)
-
-    if len(data_interp.shape) != 2:
-        raise ValueError("'data' must be a 2D array.")
-
-    if mask.shape != data.shape:
-        raise ValueError("'mask' and 'data' must have the same shape.")
-
-    y, x = np.indices(data_interp.shape)
-    xy = np.dstack((x[~mask].ravel(), y[~mask].ravel()))[0]
-    z = data_interp[~mask].ravel()
-
-    if method == 'nearest':
-        interpol = interpolate.NearestNDInterpolator(xy, z)
-    elif method == 'cubic':
-        interpol = interpolate.CloughTocher2DInterpolator(xy, z)
-    else:
-        raise ValueError('Unsupported interpolation method.')
-
-    xy_missing = np.dstack((x[mask].ravel(), y[mask].ravel()))[0]
-    data_interp[mask] = interpol(xy_missing)
-
-    return data_interp
